Revert "Pull out high-level interfaces as an extended example"
