+++ .travis.yml
@@ -31,7 +31,6 @@ jobs:
 cache: pip
 
 install:
-- pip install -U virtualenv
-- pip install tox
+- pip install tox tox-venv
 
 script: tox
