--- examples/README.rst
@@ -1,5 +1,0 @@
-Examples of using the ``pep517`` library.
-
-* ``buildtool`` is about the simplest possible PEP 517 frontend, which can use
-  PEP 517 backends to build packages. It installs build dependencies into a
-  temporary but non-isolated environment using pip.
