--- examples/buildtool/tests/samples/pkg1/pkg1-0.5.dist-info/WHEEL
@@ -1,5 +1,0 @@
-Wheel-Version: 1.0
-Generator: buildsys 0.1
-Root-Is-Purelib: true
-Tag: py2-none-any
-Tag: py3-none-any
