--- examples/buildtool/tests/samples/pkg1/pkg1.py
@@ -1,3 +1,0 @@
-"""Sample package for tests"""
-
-__version__ = '0.5'
