--- examples/buildtool/tests/samples/pkg1/pyproject.toml
@@ -1,3 +1,0 @@
-[build-system]
-requires = ["eg_buildsys"]
-build-backend = "buildsys"
