+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist = py27, py34, py35, py36, py37, pypy, pypy3
-isolated_build = True
+skipsdist = true
 
 [testenv]
 deps = -rdev-requirements.txt
